+++ src/shared/tdoc-commons/model/repository/sql-mytbdb-odimageobject.config.ts
@@ -242,6 +242,9 @@ export class SqlMytbDbOdImageObjectConfig {
                 selectFrom: 'image_object INNER JOIN image ON image_object.i_id=image.i_id',
                 orderBy: 'value asc'
             },
+            'gpstracks_state_is': {
+                noFacet: true
+            },
             'initial_s': {
                noFacet: true
             },
@@ -486,6 +489,7 @@ export class SqlMytbDbOdImageObjectConfig {
             track_id_i: 'image.k_id',
             track_id_is: 'image.k_id',
             news_id_is: '"666dummy999"',
+            gpstracks_state_is: '"666dummy999"',
             initial_s: '"666dummy999"',
             loc_lochirarchie_ids_txt: 'location.l_id',
             l_lochirarchietxt: 'location.l_name',